All changes until 04/30/2016 10:00AM
- Apple Watch data sync algorithm/functions completed & implemented
- Apple Watch basic interface finished
- Apple Watch complication implementation started
- TodayVC added
- Project CollectionView cell created & implemented
- Project VC created & implemented
- Add events to specific project & display specific events.
- AddEvent VC updated to CreateNewVC
- Small changes

Notes:
Pew, this was a lot of work. Syncing data between Apple Watch and iPhone was a big pain. But I need to constantly test this process. Now, we can add event to projects with no pain thanks to the new delegates (protocols). I have to focus on managing projects, adding gestures and cool feature. On the Apple Watch side, I have to finish complication and other small changes.