All changes until 05/11/2016
- Timeline classes added
- Storyboard timeline collectionView added
- TodayVC modified for timeline and gesture recognizer
- Material design classes added
- TableViewCells modified based on UIMaterial
- CustomCell class & other classes modified for UIMaterial
- Calendar UI and VC updated. (labels added)
- Pod files updated
- Other changes

Notes:
This is the most important commit so far in terms of UIs and designs. I have improved the other material, and also added several gesture recognizers. Timeline function is also added, as well as improvements made in Calendar UI and VC. Next, I have to make the tip controller and also intro controller. Last but not least, I have to rethink the functionalities of the main apple watch app. For example, today and timeline(maybe).