All changes until 07/09/2016
- EventManager methods tested, improved.
- Todo UI updated. Buttons added.
- TodoVC updated based on the UI changes.
- CreateNewVC modified, set some delegate methods optional.
- Made small change to code to ensure stability after analyzing the build.
- TodoVC bugs with displaying selected results are fixed.
- Other small bugs with TodayVC and TodoVC are fixed and patched.
- Pods and resource is updated.

Notes:
I have fixed and patched many small bugs in this commit. Also, I we added a button menu in TodoUI. However, this project is far from finished. First, the ATS needs to be finished. In the future, the user would be able to see all the events in a day (data from calendar) with all the tasks setup in between. The automatic task scheduling system should go with Quick Scheduling. Also, autocompletion and task tracking are needed as well. Second, I have to finished and improve some other small but critical bugs and issues. For example, the app crashes because of EXC_BAD issues, which are memory protection protocols. With a big project like this, I need to execute countless test, make countless changes as well. Last but not least, think about a timeline view. What information should I display? What information does the user want to see? Happy weekend!