All change until 07/21/2016 9:00AM
- PNChart library added.
- AnalysisVC modified, GraphKit deleted, tableView with custom cell implemented.
- GraphTableViewCell nib and class added. Some methods in the class are implemented.
- Main.storyboard modified based on the updates.
- Changes with pods.
- EventHelper -currentDateLocalTimeZone method added.

Notes:
After developing this app for such a long time, I shall not lose sight of the initial purpose and meaning of this application. Therefore, some critical feature that will expedite and improve user's experience must be added. First, easily and quickly change event date. Also, improve local notification system, some events scheduled are not notified. The automatic event management system need to be added, also display calendar events and Done tasks in different colors. Think about whether setting a project date reminder and location is necessary. If it's not, we shall simplify it to just a title, which is similar to a folder. Adding events still seems to be awkward, distill it down to date, title, that's it. Reminder is automatic, time is recommended. Once again, timeline is a great feature, however, wrong information, confusing labeling, irrelevant information will be extremely frustrating, so we must make it more intuitive. Good luck! God day!